Add files via upload
This python code can be used in the command prompt or your real computers terminal.  It will ask for the words seperated by a space.  It then evaluates the words to see which word has the most 'likeness' with the other words and suggest that word for a first guess.  If it is the correct password, 'Great!'.  Otherwise it will ask for the likeness provided by the terminal in Fallout.  Using that information it will provide the next best guess or guesses.  If there are more suggestions than guesses remaining, you can run the suggested words through the program again to narrow down the options.